Included interfaces in types.ts and polished a little
